Move Numba tensor_basic tests to test_tensor_basic
